Relocate shared  custom AttributeDefinitions in clustering/common.
