+++ clustering/common/pom.xml
@@ -49,7 +49,10 @@
         <dependency>
             <groupId>org.jboss.as</groupId>
             <artifactId>jboss-as-controller</artifactId>
-            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss</groupId>
+            <artifactId>jboss-dmr</artifactId>
         </dependency>
         <dependency>
             <groupId>org.jboss</groupId>
