+++ clustering/common/src/main/java/org/jboss/as/clustering/subsystem/validators/ObjectTypeValidator.java
@@ -20,7 +20,7 @@
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
 
-package org.jboss.as.clustering.jgroups.subsystem.validators;
+package org.jboss.as.clustering.subsystem.validators;
 
 // import static org.jboss.as.clustering.ClusteringMessages.MESSAGES;
