+++ clustering/jgroups/src/main/java/org/jboss/as/clustering/jgroups/subsystem/CommonAttributes.java
@@ -1,5 +1,8 @@
 package org.jboss.as.clustering.jgroups.subsystem;
 
+import org.jboss.as.clustering.subsystem.ObjectListAttributeDefinition;
+import org.jboss.as.clustering.subsystem.ObjectTypeAttributeDefinition;
+import org.jboss.as.clustering.subsystem.SimpleListAttributeDefinition;
 import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.SimpleAttributeDefinition;
 import org.jboss.as.controller.SimpleAttributeDefinitionBuilder;
